move the error message class to separate file
